Move Revision class and field to separate file
