+++ starknet_py/net/schemas/common.py
@@ -3,6 +3,9 @@ import sys
 from enum import Enum
 from typing import Any, Mapping, Optional, Union
 
+from starknet_py.net.schemas.revision import Revision , RevisionField
+
+
 from marshmallow import Schema, ValidationError, fields, post_load
 
 from starknet_py.net.client_models import (
@@ -348,30 +351,9 @@ class StorageEntrySchema(Schema):
         return StorageEntry(**data)
 
 
-class Revision(Enum):
-    """
-    Enum representing the revision of the specification to be used.
-    """
-
-    V0 = 0
-    V1 = 1
-
-
-class RevisionField(fields.Field):
-    def _serialize(self, value: Any, attr: Optional[str], obj: Any, **kwargs):
-        if value is None or value == Revision.V0:
-            return str(Revision.V0.value)
-        return value.value
-
-    def _deserialize(self, value, attr, data, **kwargs) -> Revision:
-        if isinstance(value, str):
-            value = int(value)
-
-        revisions = [revision.value for revision in Revision]
-        if value not in revisions:
-            allowed_revisions_str = "".join(list(map(str, revisions)))
-            raise ValidationError(
-                f"Invalid value provided for Revision: {value}. Allowed values are {allowed_revisions_str}."
-            )
-
-        return Revision(value)
+class MySchema(Schema):
+    revision = RevisionField(required=True) 
+    data = {"revision": Revision.V0, 
+            "revision": Revision.V1, 
+             
+        }
